Move to incremental generators
Moves from ISourceGenerator to IIncrementalGenerator. No significant
changes are expected, except that IDE performance should be better in
large-scale solutions.
