+++ serde-dn.sln
@@ -15,6 +15,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{C29CDEE5-5D8
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serde", "src\serde\Serde.csproj", "{2411C763-490C-4384-AD87-AD9643946F55}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serde.Generation.Test", "test\Serde.Generation.Test\Serde.Generation.Test.csproj", "{382F0F9B-7B6C-439A-96B4-BFCF433FB611}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -73,6 +75,18 @@ Global
 		{2411C763-490C-4384-AD87-AD9643946F55}.Release|x64.Build.0 = Release|Any CPU
 		{2411C763-490C-4384-AD87-AD9643946F55}.Release|x86.ActiveCfg = Release|Any CPU
 		{2411C763-490C-4384-AD87-AD9643946F55}.Release|x86.Build.0 = Release|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Debug|x64.Build.0 = Debug|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Debug|x86.Build.0 = Debug|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Release|Any CPU.Build.0 = Release|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Release|x64.ActiveCfg = Release|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Release|x64.Build.0 = Release|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Release|x86.ActiveCfg = Release|Any CPU
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -84,5 +98,6 @@ Global
 		{9050E98E-E99C-4D0B-9110-2617C65C0B49} = {278C5DCE-D96F-46AC-8865-622418F28A8D}
 		{A5391F09-58A5-4E83-9EB1-755B9CEB1812} = {278C5DCE-D96F-46AC-8865-622418F28A8D}
 		{2411C763-490C-4384-AD87-AD9643946F55} = {C29CDEE5-5D85-496E-8E43-27A474AF8AEB}
+		{382F0F9B-7B6C-439A-96B4-BFCF433FB611} = {278C5DCE-D96F-46AC-8865-622418F28A8D}
 	EndGlobalSection
 EndGlobal
