+++ src/generator/Generator.Impl.cs
@@ -1,5 +1,7 @@
 
 using System;
+using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
@@ -7,116 +9,148 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 using static Serde.Diagnostics;
 
-namespace Serde
+namespace Serde;
+
+[Flags]
+internal enum SerdeUsage : byte
 {
-    internal enum SerdeUsage : byte
+    Serialize = 0b01,
+    Deserialize = 0b10,
+
+    Both = Serialize | Deserialize,
+}
+
+internal sealed class GeneratorExecutionContext
+{
+    private bool _frozen = false;
+    private readonly List<Diagnostic> _diagnostics = new();
+    private readonly SortedSet<(string FileName, string Content)> _sources = new();
+
+    public Compilation Compilation { get; }
+
+    public GeneratorExecutionContext(GeneratorAttributeSyntaxContext context)
     {
-        Serialize = 0b01,
-        Deserialize = 0b10,
+        Compilation = context.SemanticModel.Compilation;
     }
 
-    partial class Generator
+    public GenerationOutput GetOutput()
     {
-        internal static void GenerateImpl(
-            SerdeUsage usage,
-            TypeDeclarationSyntax typeDecl,
-            SemanticModel semanticModel,
-            GeneratorExecutionContext context)
-        {
-            var receiverType = semanticModel.GetDeclaredSymbol(typeDecl);
-            if (receiverType is null)
-            {
-                return;
-            }
-            if (!typeDecl.Modifiers.Any(tok => tok.IsKind(SyntaxKind.PartialKeyword)))
-            {
-                // Type must be partial
-                context.ReportDiagnostic(CreateDiagnostic(
-                    DiagId.ERR_TypeNotPartial,
-                    typeDecl.Identifier.GetLocation(),
-                    typeDecl.Identifier.ValueText));
-                return;
-            }
+        _frozen = true;
+        return new GenerationOutput(_diagnostics, _sources);
+    }
 
-            var receiverExpr = ThisExpression();
-            GenerateImpl(usage, new TypeDeclContext(typeDecl), receiverType, receiverExpr, context);
+    public void ReportDiagnostic(Diagnostic diagnostic)
+    {
+        if (_frozen)
+            throw new InvalidOperationException("Cannot add diagnostics after GetDiagnostics() has been called.");
+        _diagnostics.Add(diagnostic);
+    }
+
+    internal void AddSource(string fileName, string content)
+    {
+        if (_frozen)
+            throw new InvalidOperationException("Cannot add sources after GetSources() has been called.");
+        _sources.Add((fileName, content));
+    }
+}
+
+partial class SerdeImplRoslynGenerator
+{
+    internal static void GenerateImpl(
+        SerdeUsage usage,
+        TypeDeclarationSyntax typeDecl,
+        SemanticModel semanticModel,
+        GeneratorExecutionContext context)
+    {
+        var receiverType = semanticModel.GetDeclaredSymbol(typeDecl);
+        if (receiverType is null)
+        {
+            return;
+        }
+        if (!typeDecl.Modifiers.Any(tok => tok.IsKind(SyntaxKind.PartialKeyword)))
+        {
+            // Type must be partial
+            context.ReportDiagnostic(CreateDiagnostic(
+                DiagId.ERR_TypeNotPartial,
+                typeDecl.Identifier.GetLocation(),
+                typeDecl.Identifier.ValueText));
+            return;
         }
 
-        private static void GenerateImpl(
-            SerdeUsage usage,
-            TypeDeclContext typeDeclContext,
-            ITypeSymbol receiverType,
-            ExpressionSyntax receiverExpr,
-            GeneratorExecutionContext context)
+        var receiverExpr = ThisExpression();
+        GenerateImpl(usage, new TypeDeclContext(typeDecl), receiverType, receiverExpr, context);
+    }
+
+    private static void GenerateImpl(
+        SerdeUsage usage,
+        TypeDeclContext typeDeclContext,
+        ITypeSymbol receiverType,
+        ExpressionSyntax receiverExpr,
+        GeneratorExecutionContext context)
+    {
+        var typeName = typeDeclContext.Name;
+        string fullTypeName = string.Join(".", typeDeclContext.NamespaceNames
+            .Concat(typeDeclContext.ParentTypeInfo.Select(x => x.Name))
+            .Concat(new[] { typeName }));
+
+        string srcName = fullTypeName + "." + usage.GetInterfaceName();
+
+        // Generate statements for the implementation
+        var (implMembers, baseList) = usage switch
         {
-            var typeName = typeDeclContext.Name;
-            string fullTypeName = string.Join(".", typeDeclContext.NamespaceNames
-                .Concat(typeDeclContext.ParentTypeInfo.Select(x => x.Name))
-                .Concat(new[] { typeName }));
+            SerdeUsage.Serialize => GenerateSerializeImpl(context, receiverType, receiverExpr),
+            SerdeUsage.Deserialize => GenerateDeserializeImpl(context, receiverType, receiverExpr),
+            _ => throw ExceptionUtilities.Unreachable
+        };
 
-            string srcName = fullTypeName + "." + usage.GetInterfaceName();
-            if (!_registry.TryAdd(context.Compilation, srcName))
-            {
-                return;
-            }
-
-            // Generate statements for the implementation
-            var (implMembers, baseList) = usage switch
-            {
-                SerdeUsage.Serialize => GenerateSerializeImpl(context, receiverType, receiverExpr),
-                SerdeUsage.Deserialize => GenerateDeserializeImpl(context, receiverType, receiverExpr),
+        var typeKind = typeDeclContext.Kind;
+        MemberDeclarationSyntax newType = TypeDeclaration(
+            typeKind,
+            attributes: default,
+            modifiers: TokenList(Token(SyntaxKind.PartialKeyword)),
+            keyword: Token(typeKind switch {
+                SyntaxKind.ClassDeclaration => SyntaxKind.ClassKeyword,
+                SyntaxKind.StructDeclaration => SyntaxKind.StructKeyword,
+                SyntaxKind.RecordDeclaration
+                or SyntaxKind.RecordStructDeclaration=> SyntaxKind.RecordKeyword,
                 _ => throw ExceptionUtilities.Unreachable
-            };
+            }),
+            identifier: Identifier(typeName),
+            typeParameterList: typeDeclContext.TypeParameterList,
+            baseList: baseList,
+            constraintClauses: default,
+            openBraceToken: Token(SyntaxKind.OpenBraceToken),
+            members: List(implMembers),
+            closeBraceToken: Token(SyntaxKind.CloseBraceToken),
+            semicolonToken: default);
 
-            var typeKind = typeDeclContext.Kind;
-            MemberDeclarationSyntax newType = TypeDeclaration(
-                typeKind,
-                attributes: default,
-                modifiers: TokenList(Token(SyntaxKind.PartialKeyword)),
-                keyword: Token(typeKind switch {
-                    SyntaxKind.ClassDeclaration => SyntaxKind.ClassKeyword,
-                    SyntaxKind.StructDeclaration => SyntaxKind.StructKeyword,
-                    SyntaxKind.RecordDeclaration
-                    or SyntaxKind.RecordStructDeclaration=> SyntaxKind.RecordKeyword,
-                    _ => throw ExceptionUtilities.Unreachable
-                }),
-                identifier: Identifier(typeName),
-                typeParameterList: typeDeclContext.TypeParameterList,
-                baseList: baseList,
-                constraintClauses: default,
-                openBraceToken: Token(SyntaxKind.OpenBraceToken),
-                members: List(implMembers),
-                closeBraceToken: Token(SyntaxKind.CloseBraceToken),
-                semicolonToken: default);
+        newType = typeDeclContext.WrapNewType(newType);
 
-            newType = typeDeclContext.WrapNewType(newType);
+        var tree = CompilationUnit(
+            externs: default,
+            usings: List(new[] { UsingDirective(IdentifierName("Serde")) }),
+            attributeLists: default,
+            members: List<MemberDeclarationSyntax>(new[] { newType }));
+        tree = tree.NormalizeWhitespace(eol: Environment.NewLine);
 
-            var tree = CompilationUnit(
-                externs: default,
-                usings: List(new[] { UsingDirective(IdentifierName("Serde")) }),
-                attributeLists: default,
-                members: List<MemberDeclarationSyntax>(new[] { newType }));
-            tree = tree.NormalizeWhitespace(eol: Environment.NewLine);
-
-            context.AddSource(srcName,
-                Environment.NewLine + "#nullable enable" + Environment.NewLine + tree.ToFullString());
-        }
+        context.AddSource(srcName,
+            Environment.NewLine + "#nullable enable" + Environment.NewLine + tree.ToFullString());
     }
+}
 
-    internal static class WrapUsageExtensions
+internal static class WrapUsageExtensions
+{
+    public static string GetInterfaceName(this SerdeUsage usage) => usage switch
     {
-        public static string GetInterfaceName(this SerdeUsage usage) => usage switch
-        {
-            SerdeUsage.Serialize => "ISerialize",
-            SerdeUsage.Deserialize => "IDeserialize",
-            _ => throw ExceptionUtilities.Unreachable
-        };
+        SerdeUsage.Serialize => "ISerialize",
+        SerdeUsage.Deserialize => "IDeserialize",
+        _ => throw ExceptionUtilities.Unreachable
+    };
 
-        public static string GetImplName(this SerdeUsage usage) => usage switch
-        {
-            SerdeUsage.Serialize => "SerializeImpl",
-            SerdeUsage.Deserialize => "DeserializeImpl",
-            _ => throw ExceptionUtilities.Unreachable
-        };
-    }
+    public static string GetImplName(this SerdeUsage usage) => usage switch
+    {
+        SerdeUsage.Serialize => "SerializeImpl",
+        SerdeUsage.Deserialize => "DeserializeImpl",
+        _ => throw ExceptionUtilities.Unreachable
+    };
 }
