+++ test/Serde.Generation.Test/JsonValueFsCheck.cs
@@ -1,5 +1,3 @@
-
-
 using System.Collections.Immutable;
 using FsCheck;
 using Serde.Json;
