+++ test/Serde.Generation.Test/MemberFormatTests.cs
@@ -88,6 +88,51 @@ enum ColorEnum
 }
 """;
             return VerifyGeneratedCode(src, new[] {
+                ("S.IDeserialize", """
+
+#nullable enable
+using Serde;
+
+partial struct S : Serde.IDeserialize<S>
+{
+    static S Serde.IDeserialize<S>.Deserialize<D>(ref D deserializer)
+    {
+        var visitor = new SerdeVisitor();
+        var fieldNames = new[]
+        {
+            "E"
+        };
+        return deserializer.DeserializeType<S, SerdeVisitor>("S", fieldNames, visitor);
+    }
+
+    private sealed class SerdeVisitor : Serde.IDeserializeVisitor<S>
+    {
+        public string ExpectedTypeName => "S";
+
+        S Serde.IDeserializeVisitor<S>.VisitDictionary<D>(ref D d)
+        {
+            Serde.Option<ColorEnum> e = default;
+            while (d.TryGetNextKey<string, StringWrap>(out string? key))
+            {
+                switch (key)
+                {
+                    case "e":
+                        e = d.GetNextValue<ColorEnum, ColorEnumWrap>();
+                        break;
+                    default:
+                        break;
+                }
+            }
+
+            var newType = new S()
+            {
+                E = e.GetValueOrThrow("E"),
+            };
+            return newType;
+        }
+    }
+}
+"""),
                 ("Serde.ColorEnumWrap", """
 
 namespace Serde
@@ -164,51 +209,6 @@ namespace Serde
                 _ when System.MemoryExtensions.SequenceEqual(s, "green"u8) => ColorEnum.Green,
                 _ when System.MemoryExtensions.SequenceEqual(s, "blue"u8) => ColorEnum.Blue,
                 _ => throw new InvalidDeserializeValueException("Unexpected enum field name: " + System.Text.Encoding.UTF8.GetString(s))};
-        }
-    }
-}
-"""),
-                ("S.IDeserialize", """
-
-#nullable enable
-using Serde;
-
-partial struct S : Serde.IDeserialize<S>
-{
-    static S Serde.IDeserialize<S>.Deserialize<D>(ref D deserializer)
-    {
-        var visitor = new SerdeVisitor();
-        var fieldNames = new[]
-        {
-            "E"
-        };
-        return deserializer.DeserializeType<S, SerdeVisitor>("S", fieldNames, visitor);
-    }
-
-    private sealed class SerdeVisitor : Serde.IDeserializeVisitor<S>
-    {
-        public string ExpectedTypeName => "S";
-
-        S Serde.IDeserializeVisitor<S>.VisitDictionary<D>(ref D d)
-        {
-            Serde.Option<ColorEnum> e = default;
-            while (d.TryGetNextKey<string, StringWrap>(out string? key))
-            {
-                switch (key)
-                {
-                    case "e":
-                        e = d.GetNextValue<ColorEnum, ColorEnumWrap>();
-                        break;
-                    default:
-                        break;
-                }
-            }
-
-            var newType = new S()
-            {
-                E = e.GetValueOrThrow("E"),
-            };
-            return newType;
         }
     }
 }
